ref: move (un)lock commands to revision folder
