+++ EMS/core-bundle/src/Command/Revision/LockCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EMS\CoreBundle\Command\ContentType;
+namespace EMS\CoreBundle\Command\Revision;
 
 use EMS\CommonBundle\Common\Command\AbstractCommand;
 use EMS\CommonBundle\Elasticsearch\Document\Document;
